child node

+++ src/Scene/Camera.cpp
@@ -105,7 +105,7 @@ void Camera::SetCameraBehavior(CameraBehavior* behavior)
 Vector3 Camera::WorldToViewportPoint(const Vector3& position) const
 {
 	const Size& size = m_ownerLayer->GetViewportSize();
-	return Vector3::Project(position, m_viewProjMatrix, 0, 0, size.Width, size.Height, m_nearClip, m_farClip);
+	return Vector3::Project(position, m_viewProjMatrix, 0.0f, 0.0f, (float)size.Width, (float)size.Height, m_nearClip, m_farClip);
 }
 
 //-----------------------------------------------------------------------------
